+++ app-client/src/index.tsx
@@ -9,36 +9,66 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import initReactFastclick from 'react-fastclick';
+import Loadable from 'react-loadable';
 import {Provider} from 'react-redux';
+import {StyleRoot} from 'radium';
 import {BrowserRouter, Route, Switch} from 'react-router-dom';
 import {ApolloClient} from 'apollo-client';
 import {ApolloProvider} from 'react-apollo';
 import {HttpLink} from 'apollo-link-http';
 import {InMemoryCache} from 'apollo-cache-inmemory';
-import {store} from './store';
-import {Demo} from './containers/demo/Demo';
+import {ToastContainer} from 'react-toastify';
+import {store} from 'src/store';
+import {LoadingPage} from 'src/containers/LoadingPage';
 
 initReactFastclick();
 
 const client = new ApolloClient({
-    link: new HttpLink({uri: 'http://localhost:4000/api'}),
+    link: new HttpLink({uri: process.env.REACT_APP_GRAPHQL_ENDPOINT}),
     cache: new InMemoryCache(),
 });
+
+// tslint:disable
+const LoadComponent = (componentName: string, path?: string) =>  {
+    return Loadable({
+        loader: () => import('src/containers/' + (path || componentName)),
+        render(loaded: any, props: any) {
+            const Component: React.ComponentClass<any> = loaded[`${componentName}`];
+
+            return <Component {...props} />;
+        },
+        loading() {
+            return <LoadingPage />;
+        }
+    });
+};
+// tslint:enable
 
 ReactDOM.render(
     <Provider store={store}>
         <ApolloProvider client={client}>
-            <div>
+            <StyleRoot>
+                <ToastContainer hideProgressBar />
                 <BrowserRouter>
                     <Switch>
+                        /*This route will be removed when the Homepage is ready*/
+                        <Route exact path="/" component={LoadComponent('Login', 'user/Login')} />
+                        <Route exact path="/login" component={LoadComponent('Login', 'user/Login')} />
+                        <Route exact path="/signup" component={LoadComponent('Signup', 'user/Signup')} />
+                        <Route exact path="/dashboard" component={LoadComponent('Dashboard', 'user/Dashboard')} />
                         <Route
-                                exact={true}
-                                path="/"
-                                component={Demo}
+                                exact
+                                path="/forgot-password"
+                                component={LoadComponent('ForgotPassword', 'user/ForgotPassword')}
                         />
+                        <Route
+                                path="/reset-password/:email"
+                                component={LoadComponent('ResetPassword', 'user/ResetPassword')}
+                        />
+                        <Route path="*" component={LoadComponent('NotFound')} />
                     </Switch>
                 </BrowserRouter>
-            </div>
+            </StyleRoot>
         </ApolloProvider>
     </Provider>,
     document.getElementById('root') as HTMLElement