+++ CHANGELOG.md
@@ -1,9 +1,19 @@
 ## [0.0.1] - 2018-01-29
 
 ### Added
+
+- Added KMS to encrypt keys.
 - RandomList generator demo.
 - Webpack loader to load graphql files.
 - Dependencies for the initial setup:
   * react-apollo
   * redux
   * graphql
+  
+## [0.0.2] - 2018-02-02
+
+### Added
+
+- Added Cognito SDKs.
+- Added Login and Signup page.
+- Added Forgot password and Reset password page.